유저 정보 입력 화면 + main 화면 slider 구현
