+++ app/src/main/AndroidManifest.xml
@@ -13,10 +13,7 @@
         android:theme="@style/Theme.SolutionChallenge"
         tools:targetApi="31">
         <activity
-            android:name=".UserEditActivity"
-            android:exported="false" />
-        <activity
-            android:name=".MainActivity"
+            android:name=".Activity.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -24,6 +21,13 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <activity
+            android:name=".Activity.UserEditActivity"
+            android:exported="false" />
+        <activity
+            android:name=".Activity.LogInActivity"
+            android:exported="false" />
+
     </application>
 
 </manifest>
