+++ app/src/main/java/com/example/solutionchallenge/Activity/LogInActivity.kt
@@ -1,0 +1,114 @@
+package com.example.solutionchallenge.Activity
+
+//import android.os.Bundle
+//import android.util.Log
+//import androidx.appcompat.app.AppCompatActivity
+//import com.google.android.gms.auth.api.signin.GoogleSignIn
+//import com.google.android.gms.auth.api.signin.GoogleSignInClient
+//import com.google.android.gms.auth.api.signin.GoogleSignInOptions
+import android.content.Intent
+import android.os.Bundle
+import android.util.Log
+import androidx.activity.result.ActivityResult
+import androidx.activity.result.ActivityResultLauncher
+import androidx.activity.result.contract.ActivityResultContracts
+import androidx.appcompat.app.AppCompatActivity
+import com.example.solutionchallenge.databinding.ActivityLogInBinding
+import com.google.android.gms.auth.api.signin.GoogleSignIn
+import com.google.android.gms.auth.api.signin.GoogleSignInClient
+import com.google.android.gms.auth.api.signin.GoogleSignInOptions
+import com.google.android.gms.common.api.ApiException
+import com.google.firebase.auth.FirebaseAuth
+import com.google.firebase.auth.FirebaseUser
+import com.google.firebase.auth.GoogleAuthProvider
+import com.google.firebase.auth.ktx.auth
+import com.google.firebase.ktx.Firebase
+
+
+class LogInActivity : AppCompatActivity() {
+
+    private val binding by lazy { ActivityLogInBinding.inflate(layoutInflater) }
+    private lateinit var auth: FirebaseAuth
+    private lateinit var mGoogleSignInClient: GoogleSignInClient
+    private lateinit var startGoogleLoginForResult: ActivityResultLauncher<Intent>
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(binding.root)
+
+        auth = Firebase.auth
+
+        googleInit()
+
+        binding.signinBtn.setOnClickListener {
+            val signInIntent = mGoogleSignInClient.signInIntent
+            startGoogleLoginForResult.launch(signInIntent)
+        }
+    }
+
+    private fun googleInit() {
+        val defaultWebClientId = "715605422298-3gqke8jgsiv83dismp2j4ovo085vdn1u.apps.googleusercontent.com"; // Android id X
+
+        val gso = GoogleSignInOptions.Builder(GoogleSignInOptions.DEFAULT_SIGN_IN)
+            .requestIdToken(defaultWebClientId)
+            .requestEmail()
+            .build()
+
+        mGoogleSignInClient = GoogleSignIn.getClient(this, gso)
+
+        startGoogleLoginForResult =
+            registerForActivityResult(ActivityResultContracts.StartActivityForResult()) { result: ActivityResult ->
+                if (result.resultCode == RESULT_OK) {
+                    result.data?.let { data ->
+
+                        val task = GoogleSignIn.getSignedInAccountFromIntent(data)
+
+                        try {
+                            // Google Sign In was successful, authenticate with Firebase
+                            val account = task.getResult(ApiException::class.java)!!
+                            Log.d(TAG, "firebaseAuthWithGoogle:" + account.id)
+                            firebaseAuthWithGoogle(account.idToken!!)
+                        } catch (e: ApiException) {
+                            // Google Sign In failed, update UI appropriately
+                            Log.w(TAG, "Google sign in failed", e)
+                        }
+                    }
+                    // Google Login Success
+                } else {
+                    Log.e(TAG, "Google Result Error ${result}")
+                }
+            }
+    }
+
+    // [START auth_with_google]
+    private fun firebaseAuthWithGoogle(idToken: String) {
+        val credential = GoogleAuthProvider.getCredential(idToken, null)
+        auth.signInWithCredential(credential)
+            .addOnCompleteListener(this) { task ->
+                if (task.isSuccessful) {
+                    // Sign in success, update UI with the signed-in user's information
+                    Log.d(TAG, "signInWithCredential:success")
+                    val user = auth.currentUser
+                    updateUI(user)
+                } else {
+                    // If sign in fails, display a message to the user.
+                    Log.w(TAG, "signInWithCredential:failure", task.exception)
+                    updateUI(null)
+                }
+            }
+    }
+
+
+    private fun updateUI(user: FirebaseUser?) {
+        // FirebaseUser 데이터에 따른 UI 작업
+        binding.userName.text = user?.displayName
+    }
+
+
+
+    companion object {
+        const val TAG = "MainActivity"
+    }
+}
+
+
