+++ app/src/main/java/com/example/solutionchallenge/PersonalFrame.kt
@@ -1,0 +1,21 @@
+package com.example.solutionchallenge
+
+
+import com.example.solutionchallenge.R
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+
+
+class PersonalFrame : Fragment() {
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        return inflater.inflate(
+            R.layout.mainframe_personal, container, false
+        ) as ViewGroup
+    }
+}
