+++ app/src/main/java/com/example/solutionchallenge/TodayFrame.kt
@@ -1,0 +1,20 @@
+package com.example.solutionchallenge
+
+import com.example.solutionchallenge.R
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+
+
+class TodayFrame : Fragment() {
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        return inflater.inflate(
+            R.layout.mainframe_today, container, false
+        ) as ViewGroup
+    }
+}
