+++ app/src/main/res/drawable/slidingcustom.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android">
+    <solid android:color="@color/white" />
+    <corners android:radius="15dp" />
+    <stroke android:color="@color/black" android:width="2dp"/>
+</shape>
