+++ app/src/main/res/layout/activity_user_edit.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:background="#ffffff"
     tools:ignore="SpeakableTextPresentCheck"
-    tools:context=".UserEditActivity">
+    tools:context=".Activity.UserEditActivity">
 
     <TextView
         android:id="@+id/userInfo"
@@ -350,11 +350,11 @@
 
     <!--tool bar 삽입-->
     <androidx.appcompat.widget.LinearLayoutCompat
-        android:id="@+id/linearLayoutCompat"
+        android:id="@+id/toolbar"
         android:layout_width="match_parent"
         android:layout_height="0dp"
         android:orientation="vertical"
-        tools:context=".MainActivity"
+        tools:context=".Activity.MainActivity"
         tools:layout_editor_absoluteX="0dp"
         tools:layout_editor_absoluteY="0dp"
         tools:ignore="MissingConstraints">
