+++ local.properties
@@ -5,4 +5,4 @@
 # For customization when using a Version Control System, please read the
 # header note.
 #Fri Feb 09 12:49:15 KST 2024
-sdk.dir=/Users/sonminseo/Library/Android/sdk
+sdk.dir=C\:\\Users\\smk19\\AppData\\Local\\Android\\Sdk
